refactor(mime): break into utils/mime.lua
